* 'feature/Q&A' of github.com:zhiyicx/thinksns-plus-android:
  (#591) 修复认证之后关闭相应页面无效的问题，我的页面刷新状态
  (#591) 更新认证状态
  (#601) 修复身份证号码验证错误的问题
  (#566) code 动态评论置顶标识位修改
  (#588) code 动态列表广告占位图
  (#602) code 修复

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/source/repository/DynamicCommentTopRepository.java
@@ -1,10 +1,8 @@
 package com.zhiyicx.thinksnsplus.data.source.repository;
 
 import com.zhiyicx.common.base.BaseJsonV2;
-import com.zhiyicx.thinksnsplus.data.beans.DynamicCommentToll;
 import com.zhiyicx.thinksnsplus.data.source.remote.DynamicClient;
 import com.zhiyicx.thinksnsplus.data.source.remote.ServiceManager;
-import com.zhiyicx.thinksnsplus.modules.dynamic.tollcomment.DynamicCommentTollContract;
 import com.zhiyicx.thinksnsplus.modules.dynamic.topdynamic_comment.DynamicCommentTopContract;
 
 import javax.inject.Inject;
@@ -30,8 +28,8 @@ public class DynamicCommentTopRepository implements DynamicCommentTopContract.Re
     }
 
     @Override
-    public Observable<BaseJsonV2<Integer>> stickTop(long feed_id,long comment_id, int amount, int day) {
-        return mDynamicClient.stickTopDynamicComment(feed_id,comment_id,amount,day)
+    public Observable<BaseJsonV2<Integer>> stickTop(long feed_id, long comment_id, double amount, int day) {
+        return mDynamicClient.stickTopDynamicComment(feed_id, comment_id, (int) amount, day)
                 .subscribeOn(Schedulers.io())
                 .unsubscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread());
